Change try_number from i64 to u32 and task_try from u16 to u32
The Airflow API returns try_number as a non-negative integer, so i64
was an incorrect type that allowed nonsensical negative values.
Downstream code was casting it to u16 with #[allow(clippy::cast_sign_loss,
clippy::cast_possible_truncation)], which could silently produce wrong
results. Using u32 at the API boundary eliminates the unsafe casts and
correctly models the domain (try numbers are always non-negative and
fit within u32).

+++ src/airflow/client/v1/log.rs
@@ -14,7 +14,7 @@ impl LogOperations for V1Client {
         dag_id: &str,
         dag_run_id: &str,
         task_id: &str,
-        task_try: u16,
+        task_try: u32,
     ) -> Result<Log> {
         let response = self
             .base_api(
